requirements.txt
python-telegram-bot==13.15
firebase-admin==6.1.0
pyrebase4==4.6.0
requests==2.28.2
python-dotenv==1.0.0
razorpay==1.3.0
python-dateutil==2.8.2
rate-limiter==1.5.2
cryptography==40.0.2